add backup folder

+++ install.sh
@@ -1,4 +1,9 @@
-rm ~/.vimrc
+if [ -f ~/.vimrc ]
+then
+    rm ~/.vimrc
+fi
 ln -s ~/.vim/init.vim ~/.vimrc
+cd ~/.vim
 git submodule update --init bundle/Vundle.vim
 vim +PluginInstall +qall
+cd -
